+++ src/Settings.h
@@ -21,9 +21,9 @@ class Settings: public QObject
 
     Q_PROPERTY(bool floatingHeader READ floatingHeader WRITE setFloatingHeader NOTIFY floatingHeaderChanged)
 
-    Q_PROPERTY(int fontSize READ fontSize WRITE setFontSize NOTIFY fontSizeChanged)
+    Q_PROPERTY(double fontSize READ fontSize WRITE setFontSize NOTIFY fontSizeChanged)
     Q_PROPERTY(QString fontName READ fontName WRITE setFontName NOTIFY fontNameChanged)
-    Q_PROPERTY(float lineSpacing READ lineSpacing WRITE setLineSpacing NOTIFY lineSpacingChanged)
+    Q_PROPERTY(double lineSpacing READ lineSpacing WRITE setLineSpacing NOTIFY lineSpacingChanged)
 
     Q_PROPERTY(int scrollPos READ scrollPos WRITE setScrollPos NOTIFY scrollPosChanged)
 
@@ -55,14 +55,14 @@ class Settings: public QObject
         bool floatingHeader() const;
         void setFloatingHeader(bool show);
 
-        int fontSize() const;
-        void setFontSize(int size);
+        double fontSize() const;
+        void setFontSize(double size);
 
         QString fontName() const;
         void setFontName(const QString& name);
 
-        float lineSpacing() const;
-        void setLineSpacing(float spacing);
+        double lineSpacing() const;
+        void setLineSpacing(double spacing);
 
         int scrollPos() const;
         void setScrollPos(int pos);
@@ -116,9 +116,9 @@ class Settings: public QObject
 
         bool _floatingHeader;
 
-        int _fontSize;
+        double _fontSize;
         QString _fontName;
-        float _lineSpacing;
+        double _lineSpacing;
 
         int _scrollPos;
